fix: ios safeview & lyric touchable
