+++ src/AzusaPlayer.tsx
@@ -1,10 +1,185 @@
-import React from 'react';
-import { usePlaybackListener } from './hooks/usePlayback';
+import React, { useEffect } from 'react';
+import { Linking, SafeAreaView, View } from 'react-native';
+import {
+  NavigationContainer,
+  DarkTheme as NavigationDarkTheme,
+  DefaultTheme as NavigationDefaultTheme,
+} from '@react-navigation/native';
+import { createDrawerNavigator } from '@react-navigation/drawer';
+import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
+import {
+  IconButton,
+  MD3DarkTheme,
+  MD3LightTheme,
+  adaptNavigationTheme,
+  Provider as PaperProvider,
+} from 'react-native-paper';
+import merge from 'deepmerge';
+import { useTranslation } from 'react-i18next';
+import { useSetupPlayer, Player } from './components/player/View';
+import Playlist from './components/playlist/View';
+import PlayerBottomPanel from './components/player/PlayerProgressControls';
+import MainBackground from './components/background/MainBackground';
+import { useNoxSetting } from './hooks/useSetting';
+import PlaylistDrawer from './components/playlists/View';
+import { ViewEnum } from './enums/View';
+import Settings from './components/setting/View';
+import DummySettings from './components/setting/DummySettings';
+import './localization/i18n';
+import AppOpenSplash from './components/background/AppOpenSplash';
+import Explore from './components/explore/ytmusic/View';
+import PIPLyricView from './components/player/PIPLyric';
+import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
-const AzusaPlayer = (props: any) => {
-  const playbackListeners = usePlaybackListener();
+const { LightTheme, DarkTheme } = adaptNavigationTheme({
+  reactNavigationLight: NavigationDefaultTheme,
+  reactNavigationDark: NavigationDarkTheme,
+});
 
-  return <>{props.children}</>;
+const CombinedDefaultTheme = merge(MD3LightTheme, LightTheme);
+const CombinedDarkTheme = merge(MD3DarkTheme, DarkTheme);
+const PlayerStyle = { backgroundColor: 'transparent' };
+
+const useSplash = (duration = 1000) => {
+  const [isReady, setIsReady] = React.useState(false);
+  useEffect(() => {
+    // wait for 1000 ms and set isReady to true
+    setTimeout(() => {
+      setIsReady(true);
+    }, duration);
+  });
+  return isReady;
+};
+
+const AzusaPlayer: React.FC = () => {
+  const { t } = useTranslation();
+  const isPlayerReady = useSetupPlayer();
+  const isSplashReady = useSplash(__DEV__ ? 1 : 2500);
+  const Drawer = createDrawerNavigator();
+  const Tab = createMaterialTopTabNavigator();
+  const playerStyle = useNoxSetting(state => state.playerStyle);
+  const defaultTheme = playerStyle.metaData.darkTheme
+    ? CombinedDarkTheme
+    : CombinedDefaultTheme;
+  const insets = useSafeAreaInsets();
+
+  function NoxPlayer() {
+    return (
+      <React.Fragment>
+        <Tab.Navigator style={PlayerStyle}>
+          <Tab.Screen
+            name={ViewEnum.PLAYER_COVER}
+            component={Player}
+            options={{ tabBarStyle: { display: 'none' } }}
+          />
+          <Tab.Screen
+            name={ViewEnum.PLAYER_PLAYLIST}
+            component={Playlist}
+            options={{ tabBarStyle: { display: 'none' } }}
+          />
+        </Tab.Navigator>
+        <PlayerBottomPanel />
+      </React.Fragment>
+    );
+  }
+
+  useEffect(() => {
+    function deepLinkHandler(data: { url: string }) {
+      console.log('deepLinkHandler', data.url);
+    }
+
+    // This event will be fired when the app is already open and the notification is clicked
+    const subscription = Linking.addEventListener('url', deepLinkHandler);
+
+    // When you launch the closed app from the notification or any other link
+    Linking.getInitialURL().then(url => console.log('getInitialURL', url));
+
+    return () => {
+      subscription.remove();
+    };
+  }, []);
+
+  if (!isPlayerReady || !isSplashReady) {
+    // TODO: this is the splash screen. get something cool from lottie.
+    return (
+      <SafeAreaView style={playerStyle.screenContainer}>
+        <AppOpenSplash />
+      </SafeAreaView>
+    );
+  }
+  // HACK: proof codewhisperer learns stackoverflow:
+  // https://stackoverflow.com/questions/54599305/how-to-set-background-image-with-react-native-and-react-navigation
+  return (
+    <MainBackground>
+      <PaperProvider
+        theme={{
+          ...defaultTheme,
+          colors: playerStyle.colors,
+        }}
+      >
+        <NavigationContainer
+          theme={{
+            ...defaultTheme,
+            colors: {
+              ...defaultTheme.colors,
+              ...playerStyle.colors,
+            },
+          }}
+        >
+          <View
+            style={{
+              flex: 1,
+              // Paddings to handle safe area
+              paddingTop: insets.top,
+              paddingLeft: insets.left,
+              paddingRight: insets.right,
+            }}
+          >
+            <Drawer.Navigator
+              initialRouteName={ViewEnum.PLAYER_HOME}
+              drawerContent={PlaylistDrawer}
+            >
+              <Drawer.Screen
+                name={ViewEnum.PLAYER_HOME}
+                options={{
+                  drawerIcon: () => <IconButton icon="ßhome-outline" />,
+                  title: String(t('appDrawer.homeScreenName')),
+                  header: () => null,
+                }}
+                component={NoxPlayer}
+              />
+              <Drawer.Screen
+                name={ViewEnum.EXPORE}
+                options={{
+                  drawerIcon: () => <IconButton icon="compass" />,
+                  title: String(t('appDrawer.exploreScreenName')),
+                }}
+                component={DummySettings}
+              />
+              <Drawer.Screen
+                name={ViewEnum.SETTINGS}
+                options={{
+                  drawerIcon: () => <IconButton icon="cog" />,
+                  title: String(t('appDrawer.settingScreenName')),
+                  header: () => null,
+                }}
+                component={Settings}
+              />
+              <Drawer.Screen
+                name={ViewEnum.LYRICS}
+                options={{
+                  drawerIcon: () => <IconButton icon="cog" />,
+                  title: String(t('appDrawer.settingScreenName')),
+                  header: () => null,
+                }}
+                component={PIPLyricView}
+              />
+            </Drawer.Navigator>
+          </View>
+        </NavigationContainer>
+      </PaperProvider>
+    </MainBackground>
+  );
 };
 
 export default AzusaPlayer;
